+++ cedar-policy-validator/src/types.rs
@@ -1535,6 +1535,10 @@ impl<'a> Effect<'a> {
     }
 }
 
+// PANIC SAFETY unit tests
+#[allow(clippy::panic)]
+// PANIC SAFETY unit tests
+#[allow(clippy::indexing_slicing)]
 #[cfg(test)]
 mod test {
     use std::collections::HashMap;